Python/requirements.txt
pandas==1.5.2
pymysql==1.0.2
pyESDL
python-dotenv=0.21.0